+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
-import Header from './components/Head/Header';
-import Footer from './components/Foot/Footer';
-import Homepage from './components/Home/Homepage';
+import Header from './components/Header/Header';
+import Footer from './components/Footer/Footer';
+import Homepage from './components/Homepage/Homepage';
 
 function App() {
   return (
